refactor(session): split layout.rs into layout, recovery, and state modules
Extract the monolithic layout.rs (1739 lines) into three focused modules:

- layout.rs: LayoutNode tree, SplitOrientation, split/remove/swap ops
- recovery.rs: PaneRecovery, PaneTarget, PaneSource, StartupStep
- state.rs: SessionState, SessionMode, WindowState, runtime metadata

All public re-exports in session/mod.rs are preserved so external
callers are unaffected. Tests are co-located with their module.

Fixes #101
